Added notes on Remote Client
Tested the ability to run a Remote Client. Just had to change the Client Network Url to use the Host IP address. This was hard coded to test, then reverted back to localhost after the tests. A GUI will now be developed to allow the user to enter the Host IP and to provide other options.